code refrectoring and connection
on this commit i have made changes to the controller code. to take the input information from textarea as well as from passwordfield of the UI. these will form part of the key and encrytion input.

the two documents(the encryted and decrrypted files) are able to be produced and updated by the code.

background image to the UI will be implemented on the next commit

currently input>process>output is now successful

tests are to be impleted.